+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * [#5977](https://github.com/rubocop-hq/rubocop/issues/5977): Remove Performance cops. ([@koic][])
+* Add auto-correction to `Naming/RescuedExceptionsVariableName`. ([@anthony-robin][])
 
 ## 0.67.2 (2019-04-05)